requirements.txt

pyss3
easyocr
pytesseract
pdf2image
Pillow
summa
language_tool_python
streamlit
